[controllers] Reworking game controller
There is now a menu with choice between
starting/continuing tournament and
displaying reports.

The remaining number of round is kept
to be able to go back in main menu and
see the reports.

The savings of tournaments and players
instance is automatic during the tournament.
We can update the rank of players at any time
during the tournament.